+++ README.md
@@ -4,7 +4,7 @@ This project contains utilities for building a trading universe and evaluating t
 
 ## Universe cache
 
-The latest selected trading universe is saved to `config/current_universe.json`. It is written by `f1_universe.update_universe()` and loaded on startup via `load_universe_from_file()`.
+The latest selected trading universe is saved to `config/current_universe.json`. It is written by `f1_universe.universe_selector.update_universe()` and loaded on startup via `load_universe_from_file()`.
 
 External processes such as `signal_loop.py` can use this file to share the same universe.
