+++ config/dynamic_params.json
@@ -1,0 +1,3 @@
+{
+  "learning_rate": 0.1
+}
