+++ config/order.json
@@ -1,0 +1,4 @@
+{
+  "max_retry": 3,
+  "default_qty": 0.0
+}
