+++ f1_universe/__init__.py
@@ -1,0 +1,1 @@
+from .universe_selector import *
