+++ f3_order/__init__.py
@@ -1,0 +1,1 @@
+from .order_executor import entry
