+++ f3_order/exception_handler.py
@@ -1,0 +1,5 @@
+"""Handle order exceptions."""
+
+def handle_exception(exc: Exception) -> None:
+    """Placeholder for exception handling."""
+    pass
