+++ f3_order/kpi_guard.py
@@ -1,0 +1,5 @@
+"""Quality assurance guard for KPIs."""
+
+def check_kpis() -> None:
+    """Placeholder KPI check."""
+    pass
