+++ f3_order/position_manager.py
@@ -1,0 +1,5 @@
+"""Manage open positions."""
+
+def monitor_positions() -> None:
+    """Placeholder for position monitoring loop."""
+    pass
