+++ f3_order/smart_buy.py
@@ -1,0 +1,5 @@
+"""Hybrid market/IOC order functions."""
+
+def smart_buy(symbol: str, amount: float) -> None:
+    """Placeholder for smart buy logic."""
+    pass
