+++ f3_order/utils.py
@@ -1,0 +1,5 @@
+"""Shared utilities for the F3 order package."""
+
+def log(message: str) -> None:
+    """Placeholder logging utility."""
+    print(message)
